gia-api/requirements/common.txt
# Django
django==2.2.13
django-bootstrap3==12.1.0
django-crispy-forms==1.9.1
django-decorator-include==3.0
django-extensions==2.2.9
django-filter==2.3.0
django-tables2==2.3.1
django-allauth==0.42.0
djmail==2.0.0

# REST
django-cors-headers==3.3.0
djangorestframework==3.11.0
djangorestframework-filters==1.0.0.dev0
django-rest-auth==0.9.5
drf-extensions==0.6.0
drf-jwt==1.16.1
drf-yasg==1.17.1

# Configuration
django-dotenv==1.4.2
django-environ==0.4.5

# PostgreSQL
psycopg2==2.8.5

# Libs for parser
requests==2.23.0
beautifulsoup4==4.9.1
lxml==4.5.1
openpyxl==3.0.3
